# Access DBImpl* and CFD* by CFHImpl* in Iterators