… db/blob/db_blob_index_test.cc
… db/db_iter.cc
… db/db_iter.h
… db/db_iterator_test.cc
… table/sst_file_reader.cc
… utilities/blob_db/blob_db_impl.cc
… utilities/transactions/write_prepared_txn_db.cc
… utilities/transactions/write_unprepared_txn_db.cc